requirements.txt
flask
flask-jwt-extended
flask-cors
python-dotenv
bcrypt